requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
pyYaml
joblib
requests
imblearn
jupyterlab
flask
python-dotenv
pymongo==4.2.0
scipy
dill
# pymongo[srv]==4.2.0 # for connecting to mongo atlas server


# mlflow
# dvc
# dvc-s3


-e .
